gradle: move texturepacker gradle plugin too new extensions directory
module: add new extensions directory

gradle: move 'tools' project too new extensions directory

tools: move bin packing logic into 'core' module and update gradle scripts to depend on core vs the other way around

bin packer: rename 'Rect' to 'BinRect' to prevent name clashing with existing 'Rect' class
