+++ core/build.gradle.kts
@@ -48,7 +48,6 @@ kotlin {
                 implementation(libs.kotlinx.atomicfu)
                 implementation(libs.kotlinx.coroutines.core)
                 implementation(libs.kotlinx.serialization.json)
-                api(project(":tools"))
             }
         }
         val commonTest by getting {
