+++ extensions/tools/build.gradle.kts
@@ -1,0 +1,56 @@
+import org.jetbrains.kotlin.gradle.plugin.KotlinJsCompilerType
+import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
+
+plugins {
+    kotlin("multiplatform")
+    kotlin("plugin.serialization")
+    id("littlekt.convention.publication")
+}
+
+repositories {
+    mavenCentral()
+}
+
+kotlin {
+    jvm {
+        compilations.all {
+            kotlinOptions.jvmTarget = "11"
+        }
+        testRuns["test"].executionTask.configure {
+            useJUnit()
+        }
+    }
+    js(KotlinJsCompilerType.IR) {
+        browser {
+            binaries.executable()
+            testTask {
+                useKarma {
+                    useChromeHeadless()
+                }
+            }
+        }
+
+        this.attributes.attribute(
+            KotlinPlatformType.attribute,
+            KotlinPlatformType.js
+        )
+
+        compilations.all {
+            kotlinOptions.sourceMap = true
+        }
+    }
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(libs.kotlinx.serialization.json)
+                implementation(project(":core"))
+            }
+        }
+        val commonTest by getting {
+            dependencies {
+                implementation(kotlin("test"))
+            }
+        }
+    }
+}
